hdmitx: Fix the coverity warning OVERFLOW_BEFORE_WIDEN
Fix the below coverity warning in
xlnx_drm_hdmi_encoder_atomic_mode_set()

Event overflow_before_widen: Potentially overflowing expression
"adjusted_mode->clock * 1000"; with type "int" (32 bits, signed) is
evaluated using 32-bit arithmetic, and then used in a context that
expects an expression of type "unsigned long" (64 bits, unsigned).

+++ hdmi/xilinx_drm_hdmi.c
@@ -1416,7 +1416,8 @@ static void xlnx_drm_hdmi_encoder_atomic_mode_set(struct drm_encoder *encoder,
 	dev_dbg(xhdmi->dev,"adjusted_mode->clock = %u Hz\n", adjusted_mode->clock);
 
 	/* request required tmds clock rate */
-	ret = clk_set_rate(xhdmi->tmds_clk, adjusted_mode->clock * 1000);
+	ret = clk_set_rate(xhdmi->tmds_clk,
+			   adjusted_mode->clock * (unsigned long)1000);
 	if (ret) {
 		dev_err(xhdmi->dev, "failed to set tmds clock rate to %d: %d\n",
 					(adjusted_mode->clock * 1000), ret);
